Feature timer (#7)
* Added create category form

Added typescript and added a model for creating categories. Started rendering the categories.

* Added category creation and deletion

Added in functionality for writing to database. Allows users to create and delete categories. Popups appear at the bottom of the screen to indicate what stage the process is at. Corrected issue of database requests being made on every render.